requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14
django-jet-reboot>=1.3
django-guardian>=2.4
drf-spectacular>=0.27
plotly>=5.18
pandas>=2.2
numpy>=1.26
asammdf>=7.0
django-environ>=0.11
python-dotenv>=1.0
psycopg[binary]>=3.1
pytest>=7.4
pytest-django>=4.5
factory-boy>=3.3
